requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.15.4
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
mysql-connector-python==9.0.0
networkx==3.1
numpy==1.24.4
openai==0.28.0
packaging==24.1
pandas==2.0.3
pillow==10.4.0
psutil==6.0.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rouge_score==0.1.2
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.1.0
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.43.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.11.1
zipp==3.19.2
